requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
lxml==5.2.2
python-docx==1.1.2
sqlparse==0.5.0
typing_extensions==4.12.0
